Don't worry, I merged it in for you humans!

+++ lib/ExpediaBookings/src/main/java/com/expedia/bookings/data/payment/checkout_payment_details.kt
@@ -32,7 +32,7 @@ data class RewardDetails(
         val paymentInstrumentId: String? = null,
         val programName: ProgramName,
         val amountToChargeInRealCurrency: Float,
-        val amountToChargeInVirtualCurrency: Int,
+        val amountToChargeInVirtualCurrency: Float,
         val rateId: String,
         val currencyCode: String)
